Consolidate types, imports
