+++ src/app/page.tsx
@@ -1,14 +1,12 @@
 'use client';
+
 //Components
 import Dashboard from '@/components/Dashboard';
 import Donate from './donate/page';
-//Libs
-import { Paper } from '@mui/material';
-//Styles
-import styles from './HomeStyles.module.css';
+import Inventory from '@/components/Inventory';
+
 //Hooks
 import { useUserContext } from '@/contexts/UserContext';
-import Inventory from '@/components/Inventory';
 
 export default function Home() {
     const currentUser = useUserContext();
